Circle fixes: extract async state init and reorder other imports
