+++ src/devtools/client/debugger/src/components/SourceOutline/isFunctionSymbol.tsx
@@ -1,0 +1,5 @@
+import { FunctionSymbol, ClassSymbol } from "../../types";
+
+export function isFunctionSymbol(symbol: FunctionSymbol | ClassSymbol): symbol is FunctionSymbol {
+  return "parameterNames" in symbol;
+}
