+++ src/devtools/client/debugger/src/utils/breakpoint/index.ts
@@ -3,7 +3,8 @@
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
 import { ThreadFront } from "protocol/thread";
-import { getBreakpoint, getSource, getSourceActorsForSource } from "../../selectors";
+import { getBreakpoint } from "../../reducers/breakpoints";
+import { getSource, getSourceActorsForSource } from "../../reducers/sources";
 import assert from "../assert";
 import { features } from "../prefs";
 import sortBy from "lodash/sortBy";
@@ -69,7 +70,7 @@ export function makeBreakpointLocation(state: UIState, location: SourceLocation)
   }
 
   let sourceUrl;
-  let sourceId;
+  let sourceId: string;
 
   if (source.url) {
     sourceUrl = source.url;
@@ -81,7 +82,7 @@ export function makeBreakpointLocation(state: UIState, location: SourceLocation)
     line: location.line,
     column: location.column,
     sourceUrl,
-    sourceId,
+    sourceId: sourceId!,
   };
 }
