+++ src/ui/setup/index.ts
@@ -14,8 +14,6 @@ import { getUserSettings } from "ui/hooks/settings";
 import { initLaunchDarkly } from "ui/utils/launchdarkly";
 import { maybeSetMixpanelContext } from "ui/utils/mixpanel";
 import { syncInitialLayoutState } from "ui/reducers/layout";
-import { getInitialTabsState } from "devtools/client/debugger/src/reducers/tabs";
-import { getInitialCommentsState } from "ui/reducers/comments";
 import { initialMessageState } from "devtools/client/webconsole/reducers/messages";
 import { trackEvent } from "ui/utils/telemetry";
 import { Recording } from "ui/types";
@@ -23,6 +21,9 @@ import { getRecording } from "ui/hooks/recordings";
 import { getRecordingId } from "ui/utils/recording";
 import { getReplaySession } from "ui/setup/prefs";
 import type { LayoutState } from "ui/state/layout";
+import type { TabsState } from "devtools/client/debugger/src/reducers/tabs";
+import { EMPTY_TABS } from "devtools/client/debugger/src/reducers/tabs";
+import { CommentsState } from "ui/state/comments";
 
 declare global {
   interface Window {
@@ -87,6 +88,30 @@ export async function getInitialLayoutState(): Promise<LayoutState> {
     showVideoPanel: "showVideoPanel" in session ? session.showVideoPanel : showVideoPanel,
     toolboxLayout: "toolboxLayout" in session ? session.toolboxLayout : toolboxLayout,
     localNags: "localNags" in session ? session.localNags : [],
+  };
+}
+
+export const getInitialTabsState = async (): Promise<TabsState> => {
+  const session = await getReplaySession(getRecordingId()!);
+
+  return { tabs: session?.tabs ?? EMPTY_TABS };
+};
+
+export async function getInitialCommentsState(): Promise<CommentsState> {
+  const recordingId = getRecordingId()!;
+
+  if (!recordingId) {
+    return {
+      hoveredComment: null,
+      pendingComment: null,
+    };
+  }
+
+  const session = await getReplaySession(recordingId);
+
+  return {
+    hoveredComment: null,
+    pendingComment: session?.pendingComment || null,
   };
 }
